+++ web/layouts/setting-layout/index.ts
@@ -1,0 +1,2 @@
+export * from "./project-setting-layout";
+export * from "./workspace-setting-layout";
