+++ web/pages/[workspaceSlug]/projects/[projectId]/settings/labels.tsx
@@ -10,7 +10,7 @@ import useUserAuth from "hooks/use-user-auth";
 import { IssueLabelService } from "services/issue";
 // layouts
 import { AppLayout } from "layouts/app-layout";
-import { ProjectSettingLayout } from "layouts/setting-layout/project-setting-layout";
+import { ProjectSettingLayout } from "layouts/setting-layout";
 // components
 import {
   CreateUpdateLabelInline,
