+++ web/pages/[workspaceSlug]/settings/billing.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 // layouts
 import { AppLayout } from "layouts/app-layout";
-import { WorkspaceSettingLayout } from "layouts/setting-layout/workspace-setting-layout";
+import { WorkspaceSettingLayout } from "layouts/setting-layout";
 // component
 import { WorkspaceSettingHeader } from "components/headers";
 // ui
